Refactoring some code a little bit.
