+++ composables/src/commonMain/kotlin/com/adammcneilly/pocketleague/composables/placeholder/PlaceholderHighlight.kt
@@ -44,8 +44,6 @@ interface PlaceholderHighlight {
 /**
  * Creates a [Fade] brush with the given initial and target colors.
  *
- * @sample com.google.accompanist.sample.placeholder.DocSample_Foundation_PlaceholderFade
- *
  * @param highlightColor the color of the highlight which is faded in/out.
  * @param animationSpec the [AnimationSpec] to configure the animation.
  */
@@ -63,8 +61,6 @@ fun PlaceholderHighlight.Companion.fade(
  * The highlight starts at the top-start, and then grows to the bottom-end during the animation.
  * During that time it is also faded in, from 0f..progressForMaxAlpha, and then faded out from
  * progressForMaxAlpha..1f.
- *
- * @sample com.google.accompanist.sample.placeholder.DocSample_Foundation_PlaceholderShimmer
  *
  * @param highlightColor the color of the highlight 'shimmer'.
  * @param animationSpec the [AnimationSpec] to configure the animation.
@@ -133,8 +129,6 @@ private data class Shimmer(
  * Creates a [PlaceholderHighlight] which fades in an appropriate color, using the
  * given [animationSpec].
  *
- * @sample com.google.accompanist.sample.placeholder.DocSample_Material_PlaceholderFade
- *
  * @param animationSpec the [AnimationSpec] to configure the animation.
  */
 @Composable
@@ -151,8 +145,6 @@ fun PlaceholderHighlight.Companion.fade(
  * The highlight starts at the top-start, and then grows to the bottom-end during the animation.
  * During that time it is also faded in, from 0f..progressForMaxAlpha, and then faded out from
  * progressForMaxAlpha..1f.
- *
- * @sample com.google.accompanist.sample.placeholder.DocSample_Material_PlaceholderShimmer
  *
  * @param animationSpec the [AnimationSpec] to configure the animation.
  * @param progressForMaxAlpha The progress where the shimmer should be at it's peak opacity.
